qoriq-atf: update build of fip_ddr for lx2

+++ recipes-bsp/atf/qoriq-atf_2.4.bb
@@ -3,14 +3,15 @@ require qoriq-atf-2.4.inc
 inherit deploy
 
 DEPENDS += "u-boot-mkimage-native u-boot openssl openssl-native mbedtls rcw cst-native bc-native"
-DEPENDS:append:lx2160a += "ddr-phy"
-DEPENDS:append:lx2162a += "ddr-phy"
 do_compile[depends] += "u-boot:do_deploy rcw:do_deploy uefi:do_deploy"
 
 PV:append = "+${SRCPV}"
 
-SRC_URI += "git://github.com/ARMmbed/mbedtls;nobranch=1;destsuffix=git/mbedtls;name=mbedtls"
+SRC_URI += "git://github.com/ARMmbed/mbedtls;nobranch=1;destsuffix=git/mbedtls;name=mbedtls \
+    git://github.com/nxp/ddr-phy-binary;nobranch=1;destsuffix=git/ddr-phy-binary;name=ddr \
+"
 SRCREV_mbedtls = "0795874acdf887290b2571b193cafd3c4041a708"
+SRCREV_ddr = "fbc036b88acb6c06ffed02c898cbae9856ec75ba"
 SRCREV_FORMAT = "atf"
 
 COMPATIBLE_MACHINE = "(qoriq)"
@@ -44,9 +45,9 @@ chassistype:ls1012a = "ls104x_1012"
 chassistype:ls1043a = "ls104x_1012"
 chassistype:ls1046a = "ls104x_1012"
 
-DDR_PHY_BIN_PATH ?= ""
-DDR_PHY_BIN_PATH:lx2160a = "${DEPLOY_DIR_IMAGE}/ddr-phy"
-DDR_PHY_BIN_PATH:lx2162a = "${DEPLOY_DIR_IMAGE}/ddr-phy"
+FIP_DDR ?= ""
+FIP_DDR:lx2160a = "${@bb.utils.contains('DISTRO_FEATURES', 'secure', 'fip_ddr', '', d)}"
+FIP_DDR:lx2162a = "${@bb.utils.contains('DISTRO_FEATURES', 'secure', 'fip_ddr', '', d)}"
 
 # requires CROSS_COMPILE set by hand as there is no configure script
 export CROSS_COMPILE="${TARGET_PREFIX}"
@@ -129,18 +130,15 @@ do_compile() {
                 cp *.pem build/${PLATFORM}/release/
             fi
 
-            oe_runmake V=1 all fip pbl PLAT=${PLATFORM} BOOT_MODE=${d} RCW=${DEPLOY_DIR_IMAGE}/rcw/${RCW_FOLDER}/${rcwimg} BL33=${UBOOT_BINARY}
+            oe_runmake V=1 all fip pbl ${FIP_DDR} PLAT=${PLATFORM} BOOT_MODE=${d} RCW=${DEPLOY_DIR_IMAGE}/rcw/${RCW_FOLDER}/${rcwimg} BL33=${UBOOT_BINARY}
             cp build/${PLATFORM}/release/bl2_${d}${SECURE_EXTENTION}.pbl .
             cp build/${PLATFORM}/release/fip.bin fip_uboot${SECURE_EXTENTION}.bin
             if [ -e build/${PLATFORM}/release/fuse_fip.bin ]; then
                 cp build/${PLATFORM}/release/fuse_fip.bin .
             fi
 
-            if [ -n "${SECURE_EXTENTION}" -a -n "${DDR_PHY_BIN_PATH}" -a ! -f ddr_fip_sec.bin ]; then
-                oe_runmake V=1 fip_ddr PLAT=${PLATFORM} DDR_PHY_BIN_PATH=${DDR_PHY_BIN_PATH}
-                if [ -e build/${PLATFORM}/release/ddr_fip_sec.bin ]; then
-                    cp build/${PLATFORM}/release/ddr_fip_sec.bin .
-                fi
+            if [ -e build/${PLATFORM}/release/ddr_fip_sec.bin ] && [ ! -f ddr_fip_sec.bin ]; then
+                cp build/${PLATFORM}/release/ddr_fip_sec.bin .
             fi
 
             if [ -e build/${PLATFORM}/release/rot_key.pem ] && [ ! -f rot_key.pem ]; then
